requirements.txt
fastapi
boto3
python-dotenv
pydantic
mangum
exceptiongroup
